web/requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
PyJWT==1.7.1